refactor(contracts): apply Rust best practices and clippy fixes
- Remove unused imports across instruction files
- Fix doc comment formatting for proper rustdoc
- Use idiomatic Rust patterns (is_ascii_lowercase, Iterator::find)
- Use enumerate for indexed loops
- Use struct initialization instead of Default::default() mutation
- Configure expected cfg values in Cargo.toml for Solana target_os
- Suppress clippy::too_many_arguments for crypto verification functions

Build now produces zero warnings with cargo build and cargo clippy.

+++ contracts/programs/zvault/src/instructions/verify_deposit.rs
@@ -224,7 +224,7 @@ pub fn process_verify_deposit(
         let tree = CommitmentTree::from_bytes_mut(&mut tree_data)?;
 
         let index = tree.next_index();
-        if index >= CommitmentTree::MAX_LEAVES as u64 {
+        if index >= CommitmentTree::MAX_LEAVES {
             return Err(ZVaultError::TreeFull.into());
         }
